evaluate | split off evaluation into its own script
* optionally plots results
* `sokrates_plot_predictions` can plot directly
