+++ setup.cfg
@@ -34,6 +34,7 @@ console_scripts =
 	eos_get_data = mlfftools.scripts.eos_get_data:app
 	# eos_get_data_from_predictions = mlfftools.scripts.eos_get_data_from_predictions:app
 	sokrates_compute = mlfftools.scripts.sokrates_compute:app
+	sokrates_evaluate = mlfftools.scripts.sokrates_evaluate:app
 	sokrates_train = mlfftools.scripts.sokrates_train:app
 	sokrates_plot_predictions = mlfftools.scripts.sokrates_plot_predictions:app
 	sokrates_plot_predictions_npz = mlfftools.scripts.sokrates_plot_predictions_npz:app
